+++ packages/amplify-category-auth/src/provider-utils/awscloudformation/service-walkthroughs/auth-questions.js
@@ -132,6 +132,10 @@ async function serviceWalkthrough(context, defaultValuesFilename, stringMapsFile
         if (answer.useDefault === 'defaultSocial') {
           coreAnswers.hostedUI = true;
         }
+
+        if (answer.useDefault === 'default') {
+          coreAnswers.hostedUI = false;
+        }
         delete answer.updateFlow;
       }
       coreAnswers = { ...coreAnswers, ...answer };