+++ packages/amplify-graphql-model-transformer/src/wrappers/object-definition-wrapper.ts
@@ -1,23 +1,23 @@
 import {
   ArgumentNode,
   DirectiveNode,
+  DocumentNode,
+  EnumTypeDefinitionNode,
   FieldDefinitionNode,
+  InputObjectTypeDefinitionNode,
   InputValueDefinitionNode,
   isListType,
+  Kind,
+  ListTypeNode,
+  Location,
+  NamedTypeNode,
   NameNode,
+  NonNullTypeNode,
   ObjectTypeDefinitionNode,
   StringValueNode,
   TypeNode,
   valueFromASTUntyped,
   ValueNode,
-  Location,
-  NonNullTypeNode,
-  ListTypeNode,
-  InputObjectTypeDefinitionNode,
-  NamedTypeNode,
-  EnumTypeDefinitionNode,
-  Kind,
-  DocumentNode,
 } from 'graphql';
 import {
   DEFAULT_SCALARS,
@@ -29,7 +29,6 @@ import {
   withNamedNodeNamed,
 } from 'graphql-transformer-common';
 
-import { merge } from 'lodash';
 // Todo: to be moved to core later. context.output.getObject would return wrapper type so its easier to manipulate
 // objects
 
@@ -73,7 +72,7 @@ export class DirectiveWrapper {
       }),
       {},
     );
-    return merge(defaultValue, argValues);
+    return Object.assign(defaultValue, argValues);
   };
 }
 
@@ -89,7 +88,7 @@ export class GenericFieldWrapper {
     this.directives = (field.directives || []).map(d => new DirectiveWrapper(d));
   }
   isList = (): boolean => {
-    return isListType(this.type);
+    return this.isListType(this.type);
   };
 
   isNonNullable = (): boolean => {
@@ -129,6 +128,15 @@ export class GenericFieldWrapper {
     }
     return false;
   };
+
+  private isListType = (type: TypeNode): boolean => {
+    if (type.kind === Kind.NON_NULL_TYPE) {
+      return isListType(type.type);
+    } else {
+      return type.kind === Kind.LIST_TYPE;
+    }
+  };
+
   public getBaseType = (): NamedTypeNode => {
     let node = this.type;
     while (node.kind === Kind.LIST_TYPE || node.kind === Kind.NON_NULL_TYPE) {
@@ -395,7 +403,7 @@ export class InputObjectDefinitionWrapper {
   };
 
   removeField = (field: InputFieldWrapper): void => {
-    if (this.hasField(field.name)) {
+    if (!this.hasField(field.name)) {
       throw new Error(`type ${this.name} does not have the field with name ${field.name}`);
     }
     const index = this.fields.indexOf(field);