+++ packages/graphql-transformers-e2e-tests/src/__tests__/SearchableWithAuthTests.e2e.test.ts
@@ -21,10 +21,11 @@ import {
   createUserPool,
   createUserPoolClient,
   addIAMRolesToCFNStack,
-  signupAndAuthenticateUser,
   createGroup,
   addUserToGroup,
   configureAmplify,
+  signupUser,
+  authenticateUser,
 } from '../cognitoUtils';
 import 'isomorphic-fetch';
 
@@ -244,15 +245,15 @@ beforeAll(async () => {
     // Configure Amplify, create users, and sign in.
     configureAmplify(USER_POOL_ID, userPoolClientId, identityPoolId);
 
-    await signupAndAuthenticateUser(USER_POOL_ID, USERNAME1, TMP_PASSWORD, REAL_PASSWORD);
-    await signupAndAuthenticateUser(USER_POOL_ID, USERNAME2, TMP_PASSWORD, REAL_PASSWORD);
-    await signupAndAuthenticateUser(USER_POOL_ID, USERNAME3, TMP_PASSWORD, REAL_PASSWORD);
+    await signupUser(USER_POOL_ID, USERNAME1, TMP_PASSWORD);
+    await signupUser(USER_POOL_ID, USERNAME2, TMP_PASSWORD);
+    await signupUser(USER_POOL_ID, USERNAME3, TMP_PASSWORD);
     await createGroup(USER_POOL_ID, WRITER_GROUP_NAME);
     await createGroup(USER_POOL_ID, ADMIN_GROUP_NAME);
     await addUserToGroup(WRITER_GROUP_NAME, USERNAME2, USER_POOL_ID);
     await addUserToGroup(ADMIN_GROUP_NAME, USERNAME2, USER_POOL_ID);
 
-    const authResAfterGroup: any = await signupAndAuthenticateUser(USER_POOL_ID, USERNAME1, TMP_PASSWORD, REAL_PASSWORD);
+    const authResAfterGroup: any = await authenticateUser(USERNAME1, TMP_PASSWORD, REAL_PASSWORD);
     const idToken = authResAfterGroup.getIdToken().getJwtToken();
     GRAPHQL_CLIENT_1 = new AWSAppSyncClient({
       url: GRAPHQL_ENDPOINT,
@@ -267,7 +268,7 @@ beforeAll(async () => {
       },
     });
 
-    const authRes2AfterGroup: any = await signupAndAuthenticateUser(USER_POOL_ID, USERNAME2, TMP_PASSWORD, REAL_PASSWORD);
+    const authRes2AfterGroup: any = await authenticateUser(USERNAME2, TMP_PASSWORD, REAL_PASSWORD);
     const idToken2 = authRes2AfterGroup.getIdToken().getJwtToken();
     GRAPHQL_CLIENT_2 = new AWSAppSyncClient({
       url: GRAPHQL_ENDPOINT,
@@ -282,7 +283,7 @@ beforeAll(async () => {
       },
     });
 
-    const authRes3: any = await signupAndAuthenticateUser(USER_POOL_ID, USERNAME3, TMP_PASSWORD, REAL_PASSWORD);
+    const authRes3: any = await authenticateUser(USERNAME3, TMP_PASSWORD, REAL_PASSWORD);
     const idToken3 = authRes3.getIdToken().getJwtToken();
     GRAPHQL_CLIENT_3 = new AWSAppSyncClient({
       url: GRAPHQL_ENDPOINT,