+++ packages/amplify-dotnet-function-template-provider/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.17](https://github.com/aws-amplify/amplify-cli/compare/amplify-dotnet-function-template-provider@1.5.16...amplify-dotnet-function-template-provider@1.5.17) (2021-08-24)
+
+**Note:** Version bump only for package amplify-dotnet-function-template-provider
+
+
+
+
+
 ## [1.5.16](https://github.com/aws-amplify/amplify-cli/compare/amplify-dotnet-function-template-provider@1.5.15...amplify-dotnet-function-template-provider@1.5.16) (2021-08-06)
 
 **Note:** Version bump only for package amplify-dotnet-function-template-provider